Move uniconfig services and workers to uniconfig packages
