+++ frinx/services/uniconfig/snapshot_manager.py
