+++ frinx/workers/uniconfig/cli_network_topology.py
@@ -1,5 +1,4 @@
 from typing import Any
-from typing import Literal
 
 from frinx.common.conductor_enums import TaskResultStatus
 from frinx.common.worker.service import ServiceWorkersImpl
@@ -9,42 +8,45 @@ from frinx.common.worker.task_def import TaskInput
 from frinx.common.worker.task_def import TaskOutput
 from frinx.common.worker.task_result import TaskResult
 from frinx.common.worker.worker import WorkerImpl
-from frinx.services.connection_manager import install_node
-from frinx.services.connection_manager import uninstall_node
+from frinx.services.uniconfig.cli_network_topology import execute
+from frinx.services.uniconfig.cli_network_topology import execute_and_read
 
 
-class ConnectionManager(ServiceWorkersImpl):
-    class InstallNode(WorkerImpl):
+class CliNetworkTopology(ServiceWorkersImpl):
+    class ExecuteAndRead(WorkerImpl):
         class WorkerDefinition(TaskDefinition):
-            name = 'Install_node_RPC'
-            description = 'Install node to Uniconfig'
+            name = 'Execute_and_read_RPC'
+            description = 'Run execute and read RPC'
 
         class WorkerInput(TaskInput):
             node_id: str
-            connection_type: Literal['netconf', 'cli']
-            install_params: dict[str, Any] | None = None,
+            command: str
+            transaction_id: str
+            wait_for_output: int = 0
             uniconfig_url_base: str | None = None
 
         class WorkerOutput(TaskOutput):
             output: dict[str, Any]
 
         def execute(self, task: Task) -> TaskResult:
-            response = install_node(**task.input_data)
+            response = execute_and_read(**task.input_data)
             return TaskResult(status=TaskResultStatus.COMPLETED, output=response.json())
 
-    class UninstallNode(WorkerImpl):
+    class Execute(WorkerImpl):
         class WorkerDefinition(TaskDefinition):
-            name = 'Uninstall_node_RPC'
-            description = 'Uninstall node from Uniconfig'
+            name = 'Execute_RPC'
+            description = 'Run execute RPC'
+            labels = ['UNICONFIG']
 
         class WorkerInput(TaskInput):
             node_id: str
-            connection_type: Literal['netconf', 'cli']
-            uniconfig_url_base: str | None = None
+            command: str
+            transaction_id: str
+            uniconfig_url_base: str | None
 
         class WorkerOutput(TaskOutput):
             output: dict[str, Any]
 
         def execute(self, task: Task) -> TaskResult:
-            response = uninstall_node(**task.input_data)
+            response = execute(**task.input_data)
             return TaskResult(status=TaskResultStatus.COMPLETED, output=response.json())
