+++ frinx/workers/uniconfig/connection_manager.py
@@ -1,4 +1,5 @@
 from typing import Any
+from typing import Literal
 
 from frinx.common.conductor_enums import TaskResultStatus
 from frinx.common.worker.service import ServiceWorkersImpl
@@ -8,45 +9,42 @@ from frinx.common.worker.task_def import TaskInput
 from frinx.common.worker.task_def import TaskOutput
 from frinx.common.worker.task_result import TaskResult
 from frinx.common.worker.worker import WorkerImpl
-from frinx.services.cli_network_topology import execute
-from frinx.services.cli_network_topology import execute_and_read
+from frinx.services.uniconfig.connection_manager import install_node
+from frinx.services.uniconfig.connection_manager import uninstall_node
 
 
-class CliNetworkTopology(ServiceWorkersImpl):
-    class ExecuteAndRead(WorkerImpl):
+class ConnectionManager(ServiceWorkersImpl):
+    class InstallNode(WorkerImpl):
         class WorkerDefinition(TaskDefinition):
-            name = 'Execute_and_read_RPC'
-            description = 'Run execute and read RPC'
+            name = 'Install_node_RPC'
+            description = 'Install node to Uniconfig'
 
         class WorkerInput(TaskInput):
             node_id: str
-            command: str
-            transaction_id: str
-            wait_for_output: int = 0
+            connection_type: Literal['netconf', 'cli']
+            install_params: dict[str, Any] | None = None,
             uniconfig_url_base: str | None = None
 
         class WorkerOutput(TaskOutput):
             output: dict[str, Any]
 
         def execute(self, task: Task) -> TaskResult:
-            response = execute_and_read(**task.input_data)
+            response = install_node(**task.input_data)
             return TaskResult(status=TaskResultStatus.COMPLETED, output=response.json())
 
-    class Execute(WorkerImpl):
+    class UninstallNode(WorkerImpl):
         class WorkerDefinition(TaskDefinition):
-            name = 'Execute_RPC'
-            description = 'Run execute RPC'
-            labels = ['UNICONFIG']
+            name = 'Uninstall_node_RPC'
+            description = 'Uninstall node from Uniconfig'
 
         class WorkerInput(TaskInput):
             node_id: str
-            command: str
-            transaction_id: str
-            uniconfig_url_base: str | None
+            connection_type: Literal['netconf', 'cli']
+            uniconfig_url_base: str | None = None
 
         class WorkerOutput(TaskOutput):
             output: dict[str, Any]
 
         def execute(self, task: Task) -> TaskResult:
-            response = execute(**task.input_data)
+            response = uninstall_node(**task.input_data)
             return TaskResult(status=TaskResultStatus.COMPLETED, output=response.json())
